# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Service Ticket Reporting Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/product-failure:
    get:
      operationId: getProductFailureReport
      tags: [reports]
      summary: Product-wise failure analysis
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: servicePartner
          in: query
          schema:
            type: ["string", "null"]
        - name: ash
          in: query
          schema:
            type: ["string", "null"]
        - name: rsh
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Product failure data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductFailureItem"
  /reports/component-failure:
    get:
      operationId: getComponentFailureReport
      tags: [reports]
      summary: Component-wise failure analysis
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: product
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Component failure data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ComponentFailureItem"
  /reports/warranty-analysis:
    get:
      operationId: getWarrantyAnalysisReport
      tags: [reports]
      summary: Warranty-wise ticket analysis (IW, OW, Extended, Jeevanam)
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: product
          in: query
          schema:
            type: ["string", "null"]
        - name: servicePartner
          in: query
          schema:
            type: ["string", "null"]
        - name: ash
          in: query
          schema:
            type: ["string", "null"]
        - name: rsh
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Warranty analysis data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WarrantyAnalysisResult"
  /reports/sales-complaint:
    get:
      operationId: getSalesComplaintReport
      tags: [reports]
      summary: Sales vs complaint (BD tickets) analysis
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: product
          in: query
          schema:
            type: ["string", "null"]
        - name: state
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Sales vs complaint data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalesComplaintItem"
  /reports/tat-analysis:
    get:
      operationId: getTatAnalysisReport
      tags: [reports]
      summary: TAT (turnaround time) analysis per activity stage
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: product
          in: query
          schema:
            type: ["string", "null"]
        - name: servicePartner
          in: query
          schema:
            type: ["string", "null"]
        - name: ash
          in: query
          schema:
            type: ["string", "null"]
        - name: rsh
          in: query
          schema:
            type: ["string", "null"]
        - name: state
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: TAT analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TatAnalysisResult"
  /reports/mrf-analysis:
    get:
      operationId: getMrfAnalysisReport
      tags: [reports]
      summary: MRF (Material Request Form) analysis and approval tracking
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
        - name: servicePartner
          in: query
          schema:
            type: ["string", "null"]
        - name: ash
          in: query
          schema:
            type: ["string", "null"]
        - name: rsh
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: MRF analysis data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MrfAnalysisResult"
...
components:
  schemas:
...
    CountByLabel:
      type: object
      required: [label, count]
      properties:
        label:
          type: string
        count:
          type: integer
...
    AvgTatByLabel:
      type: object
      required: [label, count, avgTatMinutes]
      properties:
        label:
          type: string
        count:
          type: integer
        avgTatMinutes:
          type: ["number", "null"]
...
    ProductFailureItem:
      type: object
      required: [product, category, totalTickets, breakdownTickets, pmTickets, warrantyCount, outOfWarrantyCount]
      properties:
        product:
          type: string
        category:
          type: string
        totalTickets:
          type: integer
        breakdownTickets:
          type: integer
        pmTickets:
          type: integer
        warrantyCount:
          type: integer
        outOfWarrantyCount:
          type: integer
    ComponentFailureItem:
      type: object
      required: [component, partCode, usageCount, productList]
      properties:
        component:
          type: string
        partCode:
          type: string
        usageCount:
          type: integer
        productList:
          type: array
          items:
            type: string
    WarrantyAnalysisResult:
      type: object
      required: [summary, byProduct, byServicePartner, byState]
      properties:
        summary:
          type: object
          required: [inWarranty, outOfWarranty, unverified, extended, jeevanam]
          properties:
            inWarranty:
              type: integer
            outOfWarranty:
              type: integer
            unverified:
              type: integer
            extended:
              type: integer
            jeevanam:
              type: integer
        byProduct:
          type: array
          items:
            $ref: "#/components/schemas/WarrantyByLabel"
        byServicePartner:
          type: array
          items:
            $ref: "#/components/schemas/WarrantyByLabel"
        byState:
          type: array
          items:
            $ref: "#/components/schemas/WarrantyByLabel"
    WarrantyByLabel:
      type: object
      required: [label, inWarranty, outOfWarranty, unverified]
      properties:
        label:
          type: string
        inWarranty:
          type: integer
        outOfWarranty:
          type: integer
        unverified:
          type: integer
    SalesComplaintItem:
      type: object
      required: [product, category, state, bdTickets, totalTickets]
      properties:
        product:
          type: string
        category:
          type: string
        state:
          type: string
        bdTickets:
          type: integer
        totalTickets:
          type: integer
    TatAnalysisResult:
      type: object
      required: [overall, stages, byServicePartner, byAsh, byRsh, byState]
      properties:
        overall:
          type: object
          required: [avgTatMinutes, medianTatMinutes, within24h, within48h, within72h, above72h]
          properties:
            avgTatMinutes:
              type: ["number", "null"]
            medianTatMinutes:
              type: ["number", "null"]
            within24h:
              type: integer
            within48h:
              type: integer
            within72h:
              type: integer
            above72h:
              type: integer
        stages:
          type: object
          required: [wipAvgMinutes, mrfRequestAvgMinutes, mrfApprovedAvgMinutes, dispatchedAvgMinutes]
          properties:
            wipAvgMinutes:
              type: ["number", "null"]
            mrfRequestAvgMinutes:
              type: ["number", "null"]
            mrfApprovedAvgMinutes:
              type: ["number", "null"]
            dispatchedAvgMinutes:
              type: ["number", "null"]
        byServicePartner:
          type: array
          items:
            $ref: "#/components/schemas/AvgTatByLabel"
        byAsh:
          type: array
          items:
            $ref: "#/components/schemas/AvgTatByLabel"
        byRsh:
          type: array
          items:
            $ref: "#/components/schemas/AvgTatByLabel"
        byState:
          type: array
          items:
            $ref: "#/components/schemas/AvgTatByLabel"
    MrfAnalysisResult:
      type: object
      required: [totalMrf, approved, pending, dispatched, byComponent, byServicePartner]
      properties:
        totalMrf:
          type: integer
        approved:
          type: integer
        pending:
          type: integer
        dispatched:
          type: integer
        byComponent:
          type: array
          items:
            $ref: "#/components/schemas/ComponentMrfItem"
        byServicePartner:
          type: array
          items:
            $ref: "#/components/schemas/CountByLabel"
    ComponentMrfItem:
      type: object
      required: [component, partCode, quantity, mrfCount]
      properties:
        component:
          type: string
        partCode:
          type: string
        quantity:
          type: integer
        mrfCount:
          type: integer
...