# lib/api-zod/src/generated/types/componentMrfItem.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Service Ticket Reporting Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ComponentMrfItem {
  component: string;
  partCode: string;
  quantity: number;
  mrfCount: number;
}